[core] Move constants to specific *\Ext.php classes
