--- src/DDTrace/Types.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace DDTrace\Types;
-
-const CACHE = 'cache';
-const HTTP_CLIENT = 'http';
-const WEB_SERVLET = 'web';
-const SQL = 'sql';
-
-const CASSANDRA = 'cassandra';
-const ELASTICSEARCH = 'elasticsearch';
-const MEMCACHED = 'memcached';
-const MONGO = 'mongodb';
-const REDIS = 'redis';
-
-const MESSAGE_CONSUMER = 'queue';
-const MESSAGE_PRODUCER = 'queue';
